refactor(config): rewrite ajv schema and regenerate json schema mirror for new top-level shape (resolves #1760)
